fix(chat): app updated application icons to pubication request (Issue #4317)
