+++ apps/chat/src/store/application/application.epics.ts
@@ -40,7 +40,6 @@ import {
   isEntityIdLocal,
   isMyEntity,
 } from '@/src/utils/app/id';
-import { splitEntityId } from '@/src/utils/app/shared-utils';
 import { translate } from '@/src/utils/app/translation';
 import { parseApplicationApiKey } from '@/src/utils/server/api';
 
@@ -64,7 +63,6 @@ import {
   AuthSelectors,
   ConversationsSelectors,
   ModelsSelectors,
-  PublicationSelectors,
   ShareSelectors,
 } from '@/src/store/selectors';
 
@@ -238,7 +236,7 @@ const updateApplicationEpic: AppEpic = (action$, state$) =>
 
       if (isMoved && payload.publicationUrl) {
         return of(
-          ApplicationActions.updateApplicationPublicationUrls({
+          PublicationActions.updateApplicationPublicationUrls({
             publicationUrl: payload.publicationUrl,
             oldApplicationId: payload.oldApplication.id,
             newApplicationId: updatedCustomApplication.id,
@@ -357,42 +355,6 @@ const updateApplicationEpic: AppEpic = (action$, state$) =>
             );
           }),
         ),
-      );
-    }),
-  );
-
-const updateApplicationPublicationUrlsEpic: AppEpic = (action$, state$) =>
-  action$.pipe(
-    ofType(ApplicationActions.updateApplicationPublicationUrls.type),
-    switchMap(({ payload }) => {
-      const publication = PublicationSelectors.selectPublicationByUrl(
-        state$.value,
-        payload.publicationUrl as string,
-      );
-
-      if (!publication || !publication?.resources) {
-        return EMPTY;
-      }
-
-      const { name } = splitEntityId(payload.newApplicationId);
-
-      const resources = publication.resources.map((resource) => ({
-        action: resource.action,
-        sourceUrl: resource.sourceUrl ?? '',
-        targetUrl:
-          resource.reviewUrl === payload.oldApplicationId
-            ? resource.targetUrl.split('/').slice(0, -1).concat(name).join('/')
-            : resource.targetUrl,
-      }));
-
-      return of(
-        PublicationActions.updatePublicationRequest({
-          url: publication.url,
-          dataToUpdate: {
-            targetFolder: publication.targetFolder,
-            resources,
-          },
-        }),
       );
     }),
   );
@@ -841,5 +803,4 @@ export const ApplicationEpics = combineEpics(
   enterEditModeEpic,
   exitEditModeEpic,
   setSelectedWidgetEpic,
-  updateApplicationPublicationUrlsEpic,
 );
